// RRC.UI/src/app/vehicle-mgmt/model/vehicle-details/vehicle-details.model.ts
export class VehicleDetails {
    VehicleID?:any;
    ReferenceNumber?:string;
    PlateNumber:string;
    PlateColor: string;
    VehicleMake: string;
    VehicleModel: string;
    // VehicleName: string;
    IsAlternativeVehicle: any;
    YearofManufacture: any;
    CarCompanyID: any;
    ContractNumber: any;
    ContractDuration: any;
    ContractStartDate: any;   
    ContractEndDate: any;
    VehicleRegistrationNumber: any;
    VehicleRegistrationExpiry:any;
    NextService:any;
    TyreChange:any;
    Notes:string;
    CreatedBy?:any;
    UpdatedBy?:any;
    CreatedDateTime?: any;
    UpdatedDateTime?:any;
    Status?:any;
    CreatedDepartment?: any;
    CurrentMileage?:any;
}